Move component MonitoringConfirmedCasesNumberCard from 'components/cards' to 'components/index/CardsMonitoring'
