+++ pages/cards/_card.vue
@@ -12,7 +12,7 @@ import ConfirmedCasesNumberCard from '@/components/index/CardsMonitoring/Confirm
 // モニタリング項目
 import MonitoringItemsOverviewCard from '@/components/index/CardsMonitoring/MonitoringItemsOverviewCard.vue'
 // モニタリング項目(1)新規陽性者数
-import MonitoringConfirmedCasesNumberCard from '@/components/cards/MonitoringConfirmedCasesNumberCard.vue'
+import MonitoringConfirmedCasesNumberCard from '@/components/index/CardsMonitoring/MonitoringConfirmedCasesNumberCard.vue'
 // モニタリング項目(2)#7119における発熱等相談件数
 import ConsultationAboutFeverNumberCard from '@/components/cards/ConsultationAboutFeverNumberCard.vue'
 // モニタリング項目(3)新規陽性者における接触歴等不明者数
